+++ lib/cldr/calendar/backend/month_compiler.ex
@@ -173,6 +173,7 @@ defmodule Cldr.Calendar.Compiler.Month do
       a week in week-based calendars..
 
       """
+      @impl true
       def periods_in_year(year) do
         months_in_year(year)
       end
@@ -181,6 +182,7 @@ defmodule Cldr.Calendar.Compiler.Month do
       Returns the number days in a given year.
 
       """
+      @spec days_in_year(year) :: Calendar.day()
       @impl true
 
       def days_in_year(year) do
@@ -188,10 +190,21 @@ defmodule Cldr.Calendar.Compiler.Month do
       end
 
       @doc """
+      Returns the number weeks in a given year.
+
+      """
+      @spec weeks_in_year(year) :: Calendar.week()
+      @impl true
+
+      def weeks_in_year(year) do
+        Month.weeks_in_year(year, __config__())
+      end
+
+      @doc """
       Returns how many days there are in the given year-month.
 
       """
-      @spec days_in_month(year, month) :: 28..31
+      @spec days_in_month(year, month) :: Calendar.day()
       @impl true
 
       def days_in_month(year, month) do