+++ mix/basic_week.ex
@@ -1,0 +1,6 @@
+require Cldr.Calendar.Compiler.Month
+
+# A week that starts on day one of the year
+defmodule Cldr.Calendar.BasicWeek do
+  use Cldr.Calendar.Base.Month, day: :first
+end